fix: keep Sessions packages within deletion-only file-count baseline
Fold CTR-SES root-contract and characterization surfaces into existing package files so pkg-file-count stays at the audited 24/19 counts without raising the deletion-only baseline.
